+++ forum-service/src/main/java/tn/esprit/springfever/services/interfaces/IAdService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface IAdService {
     public Ad addAd(Ad ad);
-    public Ad updateAd(int id,Ad ad);
-    public String deleteAd(int ad);
+    public Ad updateAd(Long id,Ad ad);
+    public String deleteAd(Long ad);
     public List<Ad> getAllAds();
 }
